Add files via upload
Not sure if loading the notebook here works. 
First try in using the python notebook for my experiment. 
Tried to make it as "automatic" as possible (untill the PCA), for the type of analysis I want to use it in. 
Used this code to work with a subset of the data, want to make it scalable (>100 samples).
All Python-y advises would be very welcome! ...bare with me on my "first" exploratory road into python.